requirements.txt
pytest>=7.4,<9
requests>=2.31,<3
python-dotenv>=1.0,<2
jsonschema>=4.21,<5
PyMySQL>=1.1,<2
allure-pytest>=2.13,<3
PyYAML>=6.0,<7
